Make import/create roles lib module more generic
