+++ lib/roles/createRoles.ts
@@ -1,27 +1,11 @@
 import { prisma } from 'db';
 import { Role } from '@prisma/client';
 
-export interface DiscordServerRole {
-  id: string
-  name: string
-  color: number
-  hoist: boolean
-  icon?: string
-  position: number
-  permissions: string
-  managed: boolean
-  mentionable: boolean
-  tags?: {
-    bot_id?: string
-    integration_id?: string
-  }[]
-}
+type RolesRecord = Record<string, Role | null>;
 
-type RolesRecord = Record<string, { discord: DiscordServerRole, charmverse: Role | null }>;
-
-// Create charmverse roles from discord roles
-export async function findOrCreateRolesFromDiscord (
-  discordServerRoles: DiscordServerRole[],
+// Create charmverse roles or find them from prisma to generate a final record
+export async function findOrCreateRoles (
+  serverRoles: {id: string, name: string}[],
   spaceId: string,
   userId: string,
   createRoles?: boolean
@@ -29,13 +13,13 @@ export async function findOrCreateRolesFromDiscord (
   createRoles = createRoles ?? true;
   const rolesRecord: RolesRecord = {};
   // Create all of the discord roles fist
-  for (const discordServerRole of discordServerRoles) {
+  for (const serverRole of serverRoles) {
     // Skip the @everyone role, this is assigned to all the members of the server
-    if (discordServerRole.name !== '@everyone') {
+    if (serverRole.name !== '@everyone') {
       // First check if a role with the same name already exist in the workspace
       const existingRole = await prisma.role.findFirst({
         where: {
-          name: discordServerRole.name,
+          name: serverRole.name,
           spaceId
         }
       });
@@ -48,11 +32,11 @@ export async function findOrCreateRolesFromDiscord (
           where: {
             spaceId_name: {
               spaceId,
-              name: discordServerRole.name
+              name: serverRole.name
             }
           },
           update: {
-            name: discordServerRole.name,
+            name: serverRole.name,
             space: {
               connect: {
                 id: spaceId
@@ -61,7 +45,7 @@ export async function findOrCreateRolesFromDiscord (
             createdBy: userId
           },
           create: {
-            name: discordServerRole.name,
+            name: serverRole.name,
             space: {
               connect: {
                 id: spaceId
@@ -72,10 +56,7 @@ export async function findOrCreateRolesFromDiscord (
         });
       }
 
-      rolesRecord[discordServerRole.id] = {
-        discord: discordServerRole,
-        charmverse: charmVerseRole
-      };
+      rolesRecord[serverRole.id] = charmVerseRole;
     }
   }
   return rolesRecord;
